refactor: encapsulate reactive pagination logic into separate hook
